Split errors.rs into multiple files
