+++ packages/vm/src/errors/vm_error.rs
@@ -1,6 +1,8 @@
+use snafu::Snafu;
 use std::fmt::{Debug, Display};
 
-use snafu::Snafu;
+use super::communication_error::CommunicationError;
+use super::ffi_error::FfiError;
 
 #[derive(Debug, Snafu)]
 #[non_exhaustive]
@@ -167,6 +169,23 @@ impl VmError {
     }
 }
 
+impl From<CommunicationError> for VmError {
+    fn from(communication_error: CommunicationError) -> Self {
+        VmError::CommunicationErr {
+            source: communication_error,
+        }
+    }
+}
+
+impl From<FfiError> for VmError {
+    fn from(ffi_error: FfiError) -> Self {
+        match ffi_error {
+            FfiError::OutOfGas {} => VmError::GasDepletion,
+            _ => VmError::FfiErr { source: ffi_error },
+        }
+    }
+}
+
 impl From<wasmer_runtime_core::cache::Error> for VmError {
     fn from(original: wasmer_runtime_core::cache::Error) -> Self {
         VmError::cache_err(format!("Wasmer cache error: {:?}", original))
@@ -210,179 +229,11 @@ impl From<wasmer_runtime_core::error::RuntimeError> for VmError {
     }
 }
 
-pub type VmResult<T> = core::result::Result<T, VmError>;
-
-/// An error in the communcation between contract and host. Those happen around imports and exports.
-#[derive(Debug, Snafu)]
-#[non_exhaustive]
-pub enum CommunicationError {
-    #[snafu(display(
-        "The Wasm memory address {} provided by the contract could not be dereferenced: {}",
-        offset,
-        msg
-    ))]
-    DerefErr {
-        /// the position in a Wasm linear memory
-        offset: u32,
-        msg: String,
-        backtrace: snafu::Backtrace,
-    },
-    #[snafu(display("Got an invalid value for iteration order: {}", value))]
-    InvalidOrder {
-        value: i32,
-        backtrace: snafu::Backtrace,
-    },
-    /// Whenever UTF-8 bytes cannot be decoded into a unicode string, e.g. in String::from_utf8 or str::from_utf8.
-    #[snafu(display("Cannot decode UTF8 bytes into string: {}", msg))]
-    InvalidUtf8 {
-        msg: String,
-        backtrace: snafu::Backtrace,
-    },
-    #[snafu(display("Region length too big. Got {}, limit {}", length, max_length))]
-    // Note: this only checks length, not capacity
-    RegionLengthTooBig {
-        length: usize,
-        max_length: usize,
-        backtrace: snafu::Backtrace,
-    },
-    #[snafu(display(
-        "Region length exceeds capacity. Length {}, capacity {}",
-        length,
-        capacity
-    ))]
-    RegionLengthExceedsCapacity {
-        length: u32,
-        capacity: u32,
-        backtrace: snafu::Backtrace,
-    },
-    #[snafu(display(
-        "Region exceeds address space. Offset {}, capacity {}",
-        offset,
-        capacity
-    ))]
-    RegionOutOfRange {
-        offset: u32,
-        capacity: u32,
-        backtrace: snafu::Backtrace,
-    },
-    #[snafu(display("Region too small. Got {}, required {}", size, required))]
-    RegionTooSmall {
-        size: usize,
-        required: usize,
-        backtrace: snafu::Backtrace,
-    },
-    #[snafu(display("Got a zero Wasm address"))]
-    ZeroAddress { backtrace: snafu::Backtrace },
-}
-
-impl CommunicationError {
-    pub(crate) fn deref_err<S: Into<String>>(offset: u32, msg: S) -> Self {
-        DerefErr {
-            offset,
-            msg: msg.into(),
-        }
-        .build()
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn invalid_order(value: i32) -> Self {
-        InvalidOrder { value }.build()
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn invalid_utf8<S: ToString>(msg: S) -> Self {
-        InvalidUtf8 {
-            msg: msg.to_string(),
-        }
-        .build()
-    }
-
-    pub(crate) fn region_length_too_big(length: usize, max_length: usize) -> Self {
-        RegionLengthTooBig { length, max_length }.build()
-    }
-
-    pub(crate) fn region_length_exceeds_capacity(length: u32, capacity: u32) -> Self {
-        RegionLengthExceedsCapacity { length, capacity }.build()
-    }
-
-    pub(crate) fn region_out_of_range(offset: u32, capacity: u32) -> Self {
-        RegionOutOfRange { offset, capacity }.build()
-    }
-
-    pub(crate) fn region_too_small(size: usize, required: usize) -> Self {
-        RegionTooSmall { size, required }.build()
-    }
-
-    pub(crate) fn zero_address() -> Self {
-        ZeroAddress {}.build()
-    }
-}
-
-pub type CommunicationResult<T> = core::result::Result<T, CommunicationError>;
-
-impl From<CommunicationError> for VmError {
-    fn from(communication_error: CommunicationError) -> Self {
-        VmError::CommunicationErr {
-            source: communication_error,
-        }
-    }
-}
-
-#[derive(Debug, Snafu)]
-pub enum FfiError {
-    #[snafu(display("Panic in FFI call"))]
-    ForeignPanic { backtrace: snafu::Backtrace },
-    #[snafu(display("bad argument passed to FFI"))]
-    BadArgument { backtrace: snafu::Backtrace },
-    #[snafu(display("Ran out of gas during FFI call"))]
-    OutOfGas {},
-    #[snafu(display("Error during FFI call: {}", error))]
-    Other {
-        error: String,
-        backtrace: snafu::Backtrace,
-    },
-}
-
-impl FfiError {
-    pub fn foreign_panic() -> Self {
-        ForeignPanic {}.build()
-    }
-
-    pub fn bad_argument() -> Self {
-        BadArgument {}.build()
-    }
-
-    pub fn out_of_gas() -> Self {
-        OutOfGas {}.build()
-    }
-
-    pub fn other<S>(error: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Other {
-            error: error.into(),
-        }
-        .build()
-    }
-}
-
-impl From<FfiError> for VmError {
-    fn from(ffi_error: FfiError) -> Self {
-        match ffi_error {
-            FfiError::OutOfGas {} => VmError::GasDepletion,
-            _ => VmError::FfiErr { source: ffi_error },
-        }
-    }
-}
-
-pub type FfiResult<T> = core::result::Result<T, FfiError>;
-
 #[cfg(test)]
 mod test {
     use super::*;
 
-    // VmError constructors
+    // constructors
 
     #[test]
     fn vm_error_cache_err_works() {
@@ -525,139 +376,6 @@ mod test {
         let error = VmError::write_access_denied();
         match error {
             VmError::WriteAccessDenied { .. } => {}
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    // CommunicationError constructors
-
-    #[test]
-    fn communication_error_deref_err() {
-        let error = CommunicationError::deref_err(345, "broken stuff");
-        match error {
-            CommunicationError::DerefErr { offset, msg, .. } => {
-                assert_eq!(offset, 345);
-                assert_eq!(msg, "broken stuff");
-            }
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_invalid_order() {
-        let error = CommunicationError::invalid_order(-745);
-        match error {
-            CommunicationError::InvalidOrder { value, .. } => assert_eq!(value, -745),
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_invalid_utf8() {
-        let error = CommunicationError::invalid_utf8("broken");
-        match error {
-            CommunicationError::InvalidUtf8 { msg, .. } => assert_eq!(msg, "broken"),
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_region_length_too_big_works() {
-        let error = CommunicationError::region_length_too_big(50, 20);
-        match error {
-            CommunicationError::RegionLengthTooBig {
-                length, max_length, ..
-            } => {
-                assert_eq!(length, 50);
-                assert_eq!(max_length, 20);
-            }
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_region_length_exceeds_capacity_works() {
-        let error = CommunicationError::region_length_exceeds_capacity(50, 20);
-        match error {
-            CommunicationError::RegionLengthExceedsCapacity {
-                length, capacity, ..
-            } => {
-                assert_eq!(length, 50);
-                assert_eq!(capacity, 20);
-            }
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_region_out_of_range_works() {
-        let error = CommunicationError::region_out_of_range(u32::MAX, 1);
-        match error {
-            CommunicationError::RegionOutOfRange {
-                offset, capacity, ..
-            } => {
-                assert_eq!(offset, u32::MAX);
-                assert_eq!(capacity, 1);
-            }
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_region_too_small_works() {
-        let error = CommunicationError::region_too_small(12, 33);
-        match error {
-            CommunicationError::RegionTooSmall { size, required, .. } => {
-                assert_eq!(size, 12);
-                assert_eq!(required, 33);
-            }
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn communication_error_zero_address() {
-        let error = CommunicationError::zero_address();
-        match error {
-            CommunicationError::ZeroAddress { .. } => {}
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    // FfiError constructors
-
-    #[test]
-    fn ffi_error_foreign_panic() {
-        let error = FfiError::foreign_panic();
-        match error {
-            FfiError::ForeignPanic { .. } => {}
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn ffi_error_bad_argument() {
-        let error = FfiError::bad_argument();
-        match error {
-            FfiError::BadArgument { .. } => {}
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn ffi_error_out_of_gas() {
-        let error = FfiError::out_of_gas();
-        match error {
-            FfiError::OutOfGas { .. } => {}
-            e => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn ffi_error_other() {
-        let error = FfiError::other("broken");
-        match error {
-            FfiError::Other { error, .. } => assert_eq!(error, "broken"),
             e => panic!("Unexpected error: {:?}", e),
         }
     }
